Fix #5946: Language behavior in tablet devices (#5947)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
Fix #5946
Weird app language update behaviour occured in the options screen, where
the language in the side panel sometimes did not update when the app
language selection was not changed. This is due to the presenter
originally relying on the result of
`translationController.updateAppLanguage(profileId, selection)` to
update the UI, but this is incorrect since this function's response
payload is the _previous_ selection.

This is fixed by updating the language when the selection has changed in
the viewmodel.

Additionally, it was observed that Kiswahili is in the list of audio
languages, but it is not a supported audio language. This has been
filtered out.

Tests have been added as well.


## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
|||
|--|--|
| [Swahili audio bug
fix.webm](https://github.com/user-attachments/assets/465a8e75-9ff9-4af4-bc6e-9dd5686ea64c)
| [2.1.2.3 app language overwritable and
sticky.webm](https://github.com/user-attachments/assets/00223080-495c-4c76-b381-b352c7d3cc77)
|

More tablet video recordings :
https://drive.google.com/drive/u/2/folders/1rkwYFPVLYN3wNMUgSGIzYZHf-fRK3olo