Fix #5906: Fix app language behavior when a non-admin profile is logged out (#5955)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
<!--
- Explain what your PR does. If this PR fixes an existing bug, please
include
- "Fixes #bugnum:" in the explanation so that GitHub can auto-close the
issue
  - when this PR is merged.
  -->

Fixes #5906

This PR ensures that all screens shown when no profile is logged in are
displayed using the system-wide language selected in the Android
settings.

Additionally, it addresses the issue (mentioned in
https://github.com/oppia/oppia-android/issues/5906#issuecomment-3224409093)
where accessing the Admin Controls screen from Profile Selection (after
entering the admin PIN) did not properly reflect the admin’s chosen
language. With this fix, once the admin profile is logged in via the
PIN, navigating to Home (and other screens) from Admin Controls will
correctly display content in the language set for that profile.

Note: `AdminPinActivity` still extends
`InjectableAutoLocalizedAppCompatActivity` since it is never launched
after logging out of another profile. Instead, it is only triggered in a
specific case: when the admin attempts to add the first non-admin
profile and the admin profile does not yet have a PIN set.

Per
https://github.com/oppia/oppia-android/pull/5955#discussion_r2379391490,
refactored the `AdminAuthActivity` to use profileId decorator instead of
passing the `profileId` via `AdminAuthActivityParams`.

## Screen Recordings

### Before

[Recording.webm](https://github.com/user-attachments/assets/bf66739e-9ca0-495a-8894-5f70cd70fcf8)

### After

[Recording.webm](https://github.com/user-attachments/assets/130edd3d-0a58-453c-9822-352c5cffe077)

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).