Fix #5784: Standardize audio and app language list ordering (#5988)
## Explanation
Fix #5784 

Before, the list of text and audio languages was bad sorted, so, as was
proposed, I sorted the languages in alphabetic form, excluding the
English from that sort and applying that as the first language.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [ ] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [ ] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [ ] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only

| Before (Inconsistent Sort) | After (Standardized Sort) |
| :--- | :--- |
|
[Video](https://github.com/user-attachments/assets/1f9704cc-40f1-47f5-aa65-65dc36bc53fe)
|
[Video](https://github.com/user-attachments/assets/3b8d0539-cda9-4fcd-9387-9092a8123115)
|