Fix #5962: Ensure parameters are initialized for all activities (#6005)
## Explanation
Fixes #5962

This PR introduces a force synchronization of platform parameter loading
state for all activities in the app. This guards against a very specific
edge case: if the app is killed due to a low memory situation, reopening
it will bypass the splash activity (which ensures parameter
initialization in concert with `AbstractOppiaApplication`) and lead to a
race between the application kicking off loading parameters, the
parameter states saving, and attempting to retrieve parameter state
(#5962 goes into much more depth).

In most cases the new `runBlocking` introduces a very slight thread
overhead only during early activity creation (since platform parameters
only load one time at the beginning of the app startup). All new app
cases will first ensure parameters are loaded through `SplashActivity`,
so this new `runBlocking` _only_ has work to do in specifically the low
memory case (or other similar manifestations of directly reloading the
old activity stack, but this is the only one I'm aware of). It also
doesn't kick off the actual loading, that happens earlier during
application creation. From what I can tell there's a very slim duration
between initialization kicking off and the activity being created (maybe
20ms according to my emulator), so it's expected that there's only a few
milliseconds of work, at most, to synchronize to at this point, one time
in the duration of the app lifecycle and only for this specific edge
case.

The real risk here is in the nature of `runBlocking` which could,
hypothetically, result in a deadlock. The team takes huge care in
avoiding main thread usage in the entirety of the domain layer, and
beyond that essentially never uses non-cooperative blocking (until now).
So long as that trend continues it seems very unlikely that this
scenario could ever result in a real problem. However, if it did then
either a timeout could be introduced (which might not actually guarantee
continuation depending on the nature of the deadlock), or the app
eventually gets terminated (a main thread deadlock should result in an
ANR).

The new test is an integration test specifically validating that
`InjectableAppCompatActivity`'s new logic does what it's supposed to,
and continues to. It emphasizes the latter since it's ultimately aiming
to be a regression guard against future changes (since failing to
initialize will likely result in failing tests, anyway). The new test
really has one primary test for validating functionality:
`testLaunchInjectableAppCompatActivity_withParameterLoading_doesNotThrowException`.
However, it has 3 other tests that 'bolster' this to try and guard
against future changes:
- Two tests demonstrate injecting a parameter directly and prove that
initialization is necessary to successfully inject it. This is needed to
demonstrate that the activity needs to do the same if it wants to inject
that parameter, and injecting the parameter is used later on in the test
(even though it can't trigger due to feature flags actually triggering
the crash earlier in the activity lifecycle; this acts as a guard if
those flags or how earlier they're used change in the future).
- One test tries to demonstrate the need for the activity's
initialization logic by attempting to start it when forcibly disabling
the initialization pathway. This effectively neuters the initialization
logic in the activity (in the same way as if it were commented out in
code).

The changes to `TestPlatformParameterModule` are needed to avoid the new
`runBlocking` call from deadlocking all app layer tests (since
attempting to load parameters requires
`TestCoroutineDispatchers.runCurrent()` since it uses test-coordinated
background dispatchers). It does this by leveraging Kotlin's
`Unconfined` dispatcher to essentially synchronously perform the loading
logic (with the necessary test coroutine synchronization barrier) within
the new `runBlocking` call, thus working around the specific limitation.
This entire process may end up being refined as part of addressing #5835
since the platform parameter initialization pathways for tests is
getting quite complicated and deviates heavily from production flows
(hence the need for this app layer-specific workaround).

Some miscellaneous notes:
- Note that the new interfaces don't need tests since they are not
introducing any logic or complex contract.
- The new test is what I consider an integration test and thus doesn't
have any corresponding production file that it's validating. This means
it won't count for code coverage.
- The change to `ExplorationActivityLocalTest` is due to some event
reodering that's happening. Some of the learner analytics events are
dependent on the initialization order and the changes in
`InjectableAppCompatActivity` cause parameters to be loaded a bit
earlier, so it seems to be logging one of the later events earlier. It's
unclear whether this matches production (or if the relative event
ordering even matters here), but it seems fine to make the change for
the test to pass.
- Note that an integration test is needed because it's not really viable
to directly test the abstract activity classes without a lot of
scaffolding (like `TestActivity`), so it's easier just to provide a very
targeted integration test suite and leverage `TestActivity` for the
actual `InjectableAppCompatActivity` testing.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- this is addressing a crash with no direct UI changes.